layer.conf: set PREFERRED_VERSION_llvm-native to ""
Issue: LIN1023-3418

Clang was whiltelisted in templates before. Now, it is whitelisted by
default since we have move all the whitelists out of templates. So
overrides clang-llvm usually set when template is enabled is not needed
anymore. so once this layer is added, we need to set
PREFERRED_VERSION_llvm-native to "" to avoid following warning: WARNING:
preferred version 16.0.1 of llvm-native not available (for item
llvm-native) WARNING: versions of llvm-native available: 15.0.7

(LOCAL REV: NOT UPSTREAM) - "Fix a wrlinux local patch"

Signed-off-by: Changqing Li <changqing.li@windriver.com>